% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_Orc_SMC.R
\name{adaptative_Orc_SMC}
\alias{adaptative_Orc_SMC}
\title{Function to run the Adaptive Online Rolling Controlled SMC (ORCSMC) algorithm}
\usage{
adaptative_Orc_SMC(B_max, K_min, K_max, K1, gamma, data, model, N1)
}
\arguments{
\item{B_max}{Integer. Maximum permissible block size (Bt).}

\item{K_min}{Integer. Minimum number of iterations (Kt).}

\item{K_max}{Integer. Maximum number of iterations (Kt).}

\item{K1}{Integer. Initial number of iterations at time t=1.}

\item{gamma}{Numeric. Approximate computational complexity budget (gamma).}

\item{data}{List. Contains the observation data (e.g., data$obs).}

\item{model}{List. Contains model parameters, state transition, and likelihood functions.}

\item{N1}{Integer. Initial number of particles at time t=1.}
}
\value{
A list containing the log-normalising constant estimates, filtering means, 
        forward particle systems, ESS history, and the adaptation histories for K, B, and N.
}
\description{
It automatically tunes the number of iterations (K), block size/lag length (B), 
and the number of particles (N) under a fixed computational complexity budget (gamma).
}
