% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_backward_smoothing.R
\name{compute_backward_smoothing}
\alias{compute_backward_smoothing}
\title{Function to calculate the marginal smoothing L1-error between particles}
\usage{
compute_backward_smoothing(H_list)
}
\arguments{
\item{H_list}{Information generated by the particle filter algorithm}
}
\value{
smoothed particles
}
\description{
Function to calculate the marginal smoothing L1-error between particles
}
