% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fkf.R
\name{compute_fkf}
\alias{compute_fkf}
\title{Function to compute filtering results for Gaussian linear model}
\usage{
compute_fkf(params, obs)
}
\arguments{
\item{params}{Information of the model}

\item{obs}{Observations}
}
\value{
A list contains:
fkf.obj is the filtering information
fks.obj is the smoothing information
}
\description{
Function to compute filtering results for Gaussian linear model
}
