% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_twisted_params.R
\name{compute_twisted_params}
\alias{compute_twisted_params}
\title{Function to compute the parameters of a twisting Gaussian mixture distribution}
\usage{
compute_twisted_params(params, psi)
}
\arguments{
\item{params}{List of Gaussian mixture parameters}

\item{psi}{Parameters of the twisting function}
}
\value{
List of twisted Gaussian mixture parameters
}
\description{
This function takes the parameters of a Gaussian mixture distribution and
a Gaussian twisting function and returns the parameters of the corresponding twisted distribution
obtained by multiplying the two together and renormalizing.
}
