% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood_bin.R
\name{evaluate_likelihood_bin}
\alias{evaluate_likelihood_bin}
\title{Compute log-likelihood for Binomial observations with logistic-transformed latent states}
\usage{
evaluate_likelihood_bin(x, y, M, model)
}
\arguments{
\item{x}{Hidden states}

\item{y}{Observations}

\item{M}{Number of experiments for the binomial distribution}

\item{model}{the model that used}
}
\value{
Log-likelihood
}
\description{
Compute log-likelihood for Binomial observations with logistic-transformed latent states
}
