% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_psi.R
\name{evaluate_psi}
\alias{evaluate_psi}
\title{Function to evaluate a twisting function psi}
\usage{
evaluate_psi(x, psi_pa)
}
\arguments{
\item{x}{Points to evalute the function at}

\item{psi_pa}{Parameters of twisting function}
}
\value{
Psi function evaluated at specified points
}
\description{
This function evaluates the Gaussian density of
a twisting function with the specified parameters at one or more points.
}
