% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_psi_tilde.R
\name{evaluate_psi_tilde}
\alias{evaluate_psi_tilde}
\title{Function to evaluate a twisting function psi_tilde}
\usage{
evaluate_psi_tilde(x, psi_pa, model)
}
\arguments{
\item{x}{Points to evalute the function at}

\item{psi_pa}{Parameters of twisting function}

\item{model}{List containing model parameters}
}
\value{
Twisting function psi-tilde evaluated at specified points
}
\description{
This function evaluates the one-step-ahead expectation of a twisting function with
the specified parameters at one or more points.
}
