% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_psi.R
\name{learn_psi}
\alias{learn_psi}
\title{Function to learn twisting log_psi function parameters}
\usage{
learn_psi(psi_t, H_prev, model)
}
\arguments{
\item{psi_t}{previous psi functions at each time t}

\item{H_prev}{a list of particles containing necessary generated from the simultaion process}

\item{model}{List containing model parameters}
}
\value{
Twisting log_psi function parameters
}
\description{
This function takes a collection of particle locations and calculates twisting log_psi function arguments.
}
