% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-Sum-Exp function to prevent underflow}
\usage{
log_sum_exp(log_weights)
}
\arguments{
\item{log_weights}{the log of weights of particles}
}
\value{
A list containing:
}
\description{
Log-Sum-Exp function to prevent underflow
}
