% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_to_coeff.R
\name{psi_to_coeff}
\alias{psi_to_coeff}
\title{Function to return the Gaussian parameters of psi into regression coefficients}
\usage{
psi_to_coeff(psi)
}
\arguments{
\item{psi}{Vector. Length of 2d with first d's are mean and following d's are diagnasis of covariance matrix}
}
\value{
A list of coefficient vectors of length 2d
}
\description{
Function to return the Gaussian parameters of psi into regression coefficients
}
