% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpf.R
\name{run_bpf}
\alias{run_bpf}
\title{Bootstrap Particle Filter (BPF) implementation}
\usage{
run_bpf(data, model, N)
}
\arguments{
\item{data}{A list containing the observed data, specifically an element `obs` of size [Time x d].}

\item{model}{A list containing model parameters (initial mean/cov, transition mean/cov) 
and functions for likelihood evaluation and observation simulation.}

\item{N}{The number of particles to use in the filter.}
}
\value{
A list containing the estimated log-marginal likelihood `logZ` at the final time step.
}
\description{
This function implements a standard Bootstrap Particle Filter to estimate the log-marginal 
likelihood of a Linear-Gaussian State-Space Model.
}
