% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_csmc.R
\name{run_CSMC}
\alias{run_CSMC}
\title{Controlled Sequential Monte Carlo (CSMC) - Full implementation}
\usage{
run_CSMC(data, Napf, K, model)
}
\arguments{
\item{data}{}

\item{Napf}{}

\item{model}{}
}
\description{
Controlled Sequential Monte Carlo (CSMC) - Full implementation
}
