% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psi_APF_rolling.R
\name{run_psi_APF_rolling}
\alias{run_psi_APF_rolling}
\title{Function of psi_APF}
\usage{
run_psi_APF_rolling(data, t, psi_t, H_prev, model, init)
}
\arguments{
\item{data}{List containing required data information on which to run the filter}

\item{t}{time}

\item{psi_t}{Parameters of the twisting function}

\item{H_prev}{previous information}

\item{model}{A list containing information of model}

\item{init}{Index to decide whether the algorithm is at its initialized block}
}
\value{
A list containing:
}
\description{
This function performs sampling and resampling procedures of
psi-Auxiliary Particle Filter with $/kappa$-adaptive resampling
}
