% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_obs.R
\name{sample_obs}
\alias{sample_obs}
\title{Function to sample observations}
\usage{
sample_obs(model, Time, d)
}
\arguments{
\item{model}{List containing model parameters}

\item{Time}{Total time length}

\item{d}{The dimension of observations}
}
\value{
A list containing the observation sequence
}
\description{
This function samples synthetic observations for each time step given the state space model.
}
