% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_twisted_initial.R
\name{sample_twisted_initial}
\alias{sample_twisted_initial}
\title{Function to provide samples from the twisted initial distribution}
\usage{
sample_twisted_initial(params, psi, N)
}
\arguments{
\item{params}{List of Gaussian mixture initial parameters}

\item{psi}{Parameters of the twisting function}

\item{N}{Number of samples to provide}
}
\value{
Samples from the twisted initial distribution
}
\description{
This function samples from a twisted initial distribution specified as a Gaussian mixture.
}
