% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_twisted_transition.R
\name{sample_twisted_transition}
\alias{sample_twisted_transition}
\title{Function to sample from a twisted transition distribution}
\usage{
sample_twisted_transition(x, model, psi, N)
}
\arguments{
\item{x}{Initial states that one-step-ahead of the states we want to sample}

\item{model}{List containing model parameters}

\item{psi}{Parameters of the twisting function}

\item{N}{Number of particles to sample}
}
\value{
The sample values
}
\description{
This function samples from a twisted transition distribution specified as a Gaussian mixture model.
}
