% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_observation_bin.R
\name{simulate_observation_bin}
\alias{simulate_observation_bin}
\title{Function to simulate observations for binomial distributions}
\usage{
simulate_observation_bin(x_particles, M)
}
\arguments{
\item{x_particles}{State which to evaluate observations at}

\item{M}{Number of experiments for the binomial distribution}
}
\value{
Observations generated at the specific state
}
\description{
Function to simulate observations for binomial distributions
}
