% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_observation_svm.R
\name{simulate_observation_svm}
\alias{simulate_observation_svm}
\title{Function to simulate observations for svm model}
\usage{
simulate_observation_svm(state, params)
}
\arguments{
\item{state}{State which to evaluate observations at}

\item{params}{Parameters of the observation distribution}
}
\value{
Observations generated at the specific state
}
\description{
Function to simulate observations for svm model
}
