% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Function to resample}
\usage{
resample(logW, mode = "res")
}
\arguments{
\item{logW}{Log weights}

\item{mode}{Which resampling scheme to use. Now we support multivariate resampling and residual resampling.}
}
\value{
List of particle indices
}
\description{
Carry out resampling by drawing the indices according to a specified vector of log weights
and the specified resampling scheme.
}
